+++ angular/projects/catalyst/src/lib/directives/proxies.ts
@@ -315,28 +315,28 @@ export class CatRadio {
 
 export declare interface CatRadioGroup extends Components.CatRadioGroup {
   /**
-   * Emitted when the radio group loses focus. 
+   * Emitted when the value is changed. 
    */
-  catBlur: EventEmitter<CustomEvent<FocusEvent>>;
+  catChange: EventEmitter<CustomEvent<any>>;
 
 }
 
 @ProxyCmp({
   defineCustomElementFn: undefined,
-  inputs: ['a11yLabel', 'disabled', 'labelLeft', 'name']
+  inputs: ['a11yLabel', 'disabled', 'labelLeft', 'name', 'value']
 })
 @Component({
   selector: 'cat-radio-group',
   changeDetection: ChangeDetectionStrategy.OnPush,
   template: '<ng-content></ng-content>',
-  inputs: ['a11yLabel', 'disabled', 'labelLeft', 'name']
+  inputs: ['a11yLabel', 'disabled', 'labelLeft', 'name', 'value']
 })
 export class CatRadioGroup {
   protected el: HTMLElement;
   constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
     c.detach();
     this.el = r.nativeElement;
-    proxyOutputs(this, this.el, ['catBlur']);
+    proxyOutputs(this, this.el, ['catChange']);
   }
 }
 
@@ -394,30 +394,25 @@ export declare interface CatSelect extends Components.CatSelect {
    * Emitted when scrolled to the bottom. 
    */
   catScrolledBottom: EventEmitter<CustomEvent<any>>;
-  /**
-   * Emitted when the select loses focus. 
-   */
-  catBlur: EventEmitter<CustomEvent<FocusEvent>>;
 
 }
 
 @ProxyCmp({
   defineCustomElementFn: undefined,
-  inputs: ['choices', 'disabled', 'hint', 'label', 'labelHidden', 'multiple', 'noSearch', 'placeholder', 'position', 'required', 'value'],
-  methods: ['setValue', 'setChoices', 'clearChoices', 'clearInput']
+  inputs: ['disabled', 'hint', 'items', 'label', 'labelHidden', 'multiple', 'placeholder', 'position', 'required', 'search', 'value']
 })
 @Component({
   selector: 'cat-select',
   changeDetection: ChangeDetectionStrategy.OnPush,
   template: '<ng-content></ng-content>',
-  inputs: ['choices', 'disabled', 'hint', 'label', 'labelHidden', 'multiple', 'noSearch', 'placeholder', 'position', 'required', 'value']
+  inputs: ['disabled', 'hint', 'items', 'label', 'labelHidden', 'multiple', 'placeholder', 'position', 'required', 'search', 'value']
 })
 export class CatSelect {
   protected el: HTMLElement;
   constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
     c.detach();
     this.el = r.nativeElement;
-    proxyOutputs(this, this.el, ['catChange', 'catSearch', 'catScrolledBottom', 'catBlur']);
+    proxyOutputs(this, this.el, ['catChange', 'catSearch', 'catScrolledBottom']);
   }
 }