+++ angular/projects/catalyst/src/lib/directives/radio-value-accessor.ts
@@ -5,7 +5,7 @@ import { ValueAccessor } from './value-accessor';
 
 @Directive({
   /* tslint:disable-next-line:directive-selector */
-  selector: 'cat-radio',
+  selector: 'cat-radio, cat-radio-group',
   host: {
     '(catChange)': 'handleChangeEvent($event.target.value)'
   },